requirements.txt
django==1.10.5
djangorestframework==3.6.3
markdown==2.6.9
django-filter==1.0.4
